requirements.txt
genanki==0.13.1
markdown==3.5.1
python-frontmatter==1.0.0
requests==2.31.0 